pipe poiseuille v2
