+++ Python/symbolic_tools/Benchmarks/HeatTransferInMultilayerPipe/plot2D_multilayer_pipe_anal_Dirichlet.py
@@ -1,16 +1,21 @@
 
 import numpy as np
 import matplotlib.pyplot as plt
-from Benchmarks.MultilayerPipe.steady_two_layer_cylinder_analytical_2D import PipeWithinPipeNeumann
-gauge = 4
+from Benchmarks.HeatTransferInMultilayerPipe.steady_two_layer_cylinder_analytical_2D import PipeWithinPipeDirichlet
+gauge = 1
 
 r0 = gauge * (8 / 2)  # inner radius
 r2 = gauge * (30 / 2)  # outer radius
+r1 = gauge * (20 / 2)  # interface between layers
 
-J0 = 1  # heat flux (dT/dr) for r = r0
-T2 = 0  # temperature for r = r2
+k1 = 0.1  # inner layer - heat conductivity for r0 < r < r1
+k2 = 0.01  # outer layer - heat conductivity for r1 < r < r2
 
-pwp = PipeWithinPipeNeumann(r0, r2, J0, T2)
+T0 = 0  # temperature for r = r0
+T2 = 1  # temperature for r = r2
+
+
+pwp = PipeWithinPipeDirichlet(r0, r1, r2, k1, k2, T0, T2)
 
 step = 0.01
 
@@ -18,7 +23,7 @@ r = np.arange(r0, r2, step)
 y = np.array([pwp.get_temperature_r(r_) for r_ in r])
 
 
-fig_name = f'pipe_within_pipe_J0{J0}_T2{T2}.png'
+fig_name = f'pipe_within_pipe_k1{k1}_k2{k2}.png'
 
 # -------------------- make dummy plot --------------------
 plt.rcParams.update({'font.size': 14})
@@ -32,8 +37,7 @@ plt.plot(r, y,
 # ------ format y axis ------ #
 yll = y.min()
 yhl = y.max()
-# axes.set_ylim([yll, yhl])
-axes.set_ylim([yll, 0.001])
+axes.set_ylim([yll, yhl])
 # axes.set_yticks(np.linspace(yll, yhl, 5))
 # axes.set_yticks(np.arange(yll, yhl, 1E-2))
 # axes.set_yticks([1E-4, 1E-6, 1E-8, 1E-10, 1E-12])
