+++ Python/symbolic_tools/Benchmarks/HeatTransferInMultilayerPipe/plot2D_multilayer_pipe_anal_vs_num_Dirichlet.py
@@ -1,4 +1,4 @@
-from Benchmarks.MultilayerPipe.steady_two_layer_cylinder_analytical_2D import PipeWithinPipeNeumann
+from Benchmarks.HeatTransferInMultilayerPipe.steady_two_layer_cylinder_analytical_2D import PipeWithinPipeDirichlet
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 from matplotlib import cm as colormap
@@ -14,51 +14,47 @@ wd = os.getcwd()
 wd = os.path.dirname(wd)  # go level up
 
 reference_lattice_size = 32
-gauge = 4
+gauge = 8
 lattice_size = int(gauge * reference_lattice_size)
 
-k = f"0.1666666"
-home = pwd.getpwuid(os.getuid()).pw_dir
-# k_0.1666666_size_128lu
 
-filename_vtk = f'neumann_bc_k_{k}_size_{int(gauge * reference_lattice_size)}lu_VTK_P00_00010000.vti'
-main_folder = os.path.join(home, 'DATA_FOR_PLOTS', 'batch_ruraWrurze_NeumannBC', f'k_{k}_size_{lattice_size}lu_bb')
+home = pwd.getpwuid(os.getuid()).pw_dir
+
+k_inner = 0.1  # inner layer - heat conductivity for r0 < r < r1
+k_outer = 0.1  # outer layer - heat conductivity for r1 < r < r2
+
+# bc_scheme = 'abb_scheme'
+bc_scheme = 'eq_scheme'
+
+filename_vtk = f'k_outer_{k_outer}_size_{int(gauge * reference_lattice_size)}lu_VTK_P00_00500000.vti'
+main_folder = os.path.join(home, 'DATA_FOR_PLOTS', 'batch_ruraWrurze_variable_k', bc_scheme, f'k_outer_{k_outer}_size_{lattice_size}lu')
 filepath_vtk = os.path.join(main_folder, filename_vtk)
 vti_reader = VTIFile(filepath_vtk)
 T_num = vti_reader.get("T")
-T_num_slice = T_num[:, :, 1]
+
 
 # ----------------------- calc dimensions -----------------------
 
-ySIZE, xSIZE = T_num_slice.shape
+ySIZE, xSIZE = T_num.shape
 assert ySIZE == xSIZE == int(gauge * reference_lattice_size)
 
 r0 = gauge * (8 / 2)  # inner radius
 r2 = gauge * (30 / 2)  # outer radius
 
-# J0 = 0.22  # heat flux (dT/dr) for r = r0
-J0 = 1  # heat flux (dT/dr) for r = r0
-T2 = 0  # temperature for r = r2
+abb_correction = 0.5
+if bc_scheme =='abb_scheme':
+    r0 += abb_correction
+    r2 -= abb_correction
 
-# ----------------------- compute anal solution ---------------------------
+r1 = gauge * (20 / 2)  # interface between layers
+
 x0 = gauge * (reference_lattice_size / 2)  # center of the pipe
 y0 = gauge * (reference_lattice_size / 2)
 
-pwp = PipeWithinPipeNeumann(r0, r2, J0, T2)
 
+# ----------------------- compute anal solution ---------------------------
 
-class Plate1DNeumann:
-    def __init__(self, J, T0):
-        """
-        :param J: heat flux
-        :param T0: temperature at x = 0
-        """
-        self.J = J
-        self.T0 = T0
-        
-    def get_temperature(self, x):
-        return self.J*x + self.T0
-
+pwp = PipeWithinPipeDirichlet(r0, r1, r2, k_inner, k_outer, T0=0, T2=1)
 
 x_grid = np.linspace(0, xSIZE, xSIZE, endpoint=False) + 0.5
 y_grid = np.linspace(0, ySIZE, ySIZE, endpoint=False) + 0.5
@@ -66,12 +62,14 @@ xx, yy = np.meshgrid(x_grid, y_grid)
 T_anal = np.zeros((ySIZE, xSIZE))
 
 for i in range(ySIZE):
+    # print(f"=== Doing i/ny: {i}/{ny}  ===")
     for j in range(xSIZE):
+        # print(f"Doing i/ny: {i}/{ny} \t j/nx: {j}/{nx}")
         r = pwp.get_r_from_xy(xx[i][j], yy[i][j], x0, y0)
         T_anal[i][j] = pwp.get_temperature_r(r)
 
 
-T_err_field_eq = T_anal - T_num_slice
+T_err_field_eq = T_anal - T_num
 # T_L2 = np.sum(abs(T_err_field)
 
 # -------- error norm hacks ---------------
@@ -81,49 +79,37 @@ T_err_field_eq = T_anal - T_num_slice
 # T_err_field = np.clip(T_err_field, -1, 1)
 # np.isnat()
 
-T_mse_eq = np.sum((T_anal - T_num_slice) * (T_anal - T_num_slice)) / len(T_anal)
+# 2d clip
+T_anal = T_anal[:, int(xSIZE / 2)]
+T_num = T_num[:, int(xSIZE / 2)]
+T_mse_eq = np.sum((T_anal - T_num) * (T_anal - T_num)) / len(T_anal)
 T_L2_eq = np.sqrt(
-    np.sum((T_anal - T_num_slice) * (T_anal - T_num_slice))
+    np.sum((T_anal - T_num) * (T_anal - T_num))
     / np.sum(T_anal * T_anal))  # Eq. 4.57
-
-x_slice = np.arange(0, int(xSIZE / 2), 1) + 0.5
-T_num_slice = T_num_slice[:, int(xSIZE / 2)]  # take Y slice
-T_num_slice = T_num_slice[int(xSIZE / 2):]  # half of it
-
-T_anal = T_anal[:, int(xSIZE / 2)]  # take Y slice
-T_anal = T_anal[int(xSIZE / 2):]  # half of it
-
-x = x_grid[:int(xSIZE / 2)]  # half of it
-
-# step = 0.01
-# r = np.arange(r0, r2, step) + 0.5
-mask = (x > r0) & (x < r2)
-r = x[mask]
-T_r_anal = np.array([pwp.get_temperature_r(r_) for r_ in r])
 
 ###################################################################################################################
 
-fig_name = f'pipe_within_pipe_Neumann_anal_vs_num_J0{J0}_T2{T2}_{xSIZE}lu.png'
+fig_name = f'pipe_within_pipe_k1_{k_inner}_k2_{k_outer}_{xSIZE}lu.png'
+
 
 # -------------------- make dummy plot --------------------
 plt.rcParams.update({'font.size': 14})
 plt.figure(figsize=(14, 8))
 
 axes = plt.gca()
-
-plt.plot(r, T_r_anal,
-         color="black", marker="", markevery=5, markersize=5, linestyle="-", linewidth=2,
+plt.plot(x_grid, T_anal,
+         color="black", marker="", markevery=1, markersize=15, linestyle="-", linewidth=2,
          label='analytical solution')
 
-plt.plot(x, T_num_slice,
+plt.plot(x_grid, T_num,
          color="black", marker="", markevery=1, markersize=15, linestyle=":", linewidth=2,
          label='current model')
 
 
 # ------ format y axis ------ #
-yll = T_r_anal.min()
-# yhl = T_r_anal.max()
-axes.set_ylim([yll, 1.05*T2])
+yll = T_anal.min()
+yhl = T_anal.max()
+axes.set_ylim([yll, 1.05*yhl])
 # axes.set_yticks(np.linspace(yll, yhl, 5))
 # axes.set_yticks(np.arange(yll, yhl, 1E-2))
 # axes.set_yticks([1E-4, 1E-6, 1E-8, 1E-10, 1E-12])
@@ -132,13 +118,14 @@ axes.set_ylim([yll, 1.05*T2])
 # plt.yscale('log')
 # ------ format x axis ------ #
 plt.xlim(0, int(xSIZE / 2))
-# plt.xlim(int(xSIZE / 2), xSIZE)
 
 # plt.ticklabel_format(style='sci', axis='x', scilimits=(0, 0))
 # plt.ticklabel_format(style='sci', axis='y', scilimits=(0, 0))  # scilimits=(-8, 8)
 
 
-plt.title(f'T across pipes,  \n '
+plt.title(f'T across pipes, {bc_scheme} \n '
+          r'$k_{inner}$=' + f'{k_inner} \t;\t' 
+          r'$k_{outer}$=' + f'{k_outer} \n'
           f'{xSIZE}x{xSIZE} [lu]'
           f'\t' r'$T_{MSE}$=' + f'{T_mse_eq:.2e}'
           )
