+++ Python/symbolic_tools/Benchmarks/HeatTransferInMultilayerPipe/plot3D_multilayer_pipe_anal_vs_num_Dirichlet.py
@@ -1,4 +1,4 @@
-from Benchmarks.MultilayerPipe.steady_two_layer_cylinder_analytical_2D import PipeWithinPipeDirichlet
+from Benchmarks.HeatTransferInMultilayerPipe.steady_two_layer_cylinder_analytical_2D import PipeWithinPipeDirichlet
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 from matplotlib.ticker import FormatStrFormatter
